Move signature_verification into submodule
Moved to avoid side effects of loading conda.trust. This will make it easier to add more trust-related functionality in the future.
